# CleanUriSegment optimization.