--- a/src/Datadog.Trace/Util/UriHelpers.cs
+++ b/src/Datadog.Trace/Util/UriHelpers.cs
@@ -6,8 +6,6 @@ namespace Datadog.Trace.Util
 {
     internal static class UriHelpers
     {
-        public const string UrlIdPlaceholder = "?";
-
         public static string CleanUri(Uri uri, bool removeScheme, bool tryRemoveIds)
         {
             var path = GetRelativeUrl(uri, tryRemoveIds);
@@ -42,8 +40,48 @@ namespace Datadog.Trace.Util
                 return absolutePath;
             }
 
+#if NETCOREAPP
+            ReadOnlySpan<char> absPath = absolutePath.AsSpan();
+
             // Sanitized url will be at worse as long as the original
-            var sb = new StringBuilder(absolutePath.Length);
+            var sb = StringBuilderCache.Acquire(absPath.Length);
+
+            int previousIndex = 0;
+            int index;
+
+            do
+            {
+                ReadOnlySpan<char> nStart = absPath.Slice(previousIndex);
+                index = nStart.IndexOf('/');
+                ReadOnlySpan<char> segment = index == -1 ? nStart : nStart.Slice(0, index);
+
+                // replace path segments that look like numbers or guid
+                // GUID format N "d85b1407351d4694939203acc5870eb1" length: 32
+                // GUID format D "d85b1407-351d-4694-9392-03acc5870eb1" length: 36 with dashes in indices 8, 13, 18 and 23.
+                if (long.TryParse(segment, out _) ||
+                    (segment.Length == 32 && IsAGuid(segment, "N")) ||
+                    (segment.Length == 36 && IsAGuid(segment, "D")))
+                {
+                    sb.Append('?');
+                }
+                else
+                {
+                    sb.Append(segment);
+                }
+
+                if (index != -1)
+                {
+                    sb.Append('/');
+                }
+
+                previousIndex += index + 1;
+            }
+            while (index != -1);
+
+            return StringBuilderCache.GetStringAndRelease(sb);
+#else
+            // Sanitized url will be at worse as long as the original
+            var sb = StringBuilderCache.Acquire(absolutePath.Length);
 
             int previousIndex = 0;
             int index = 0;
@@ -71,7 +109,7 @@ namespace Datadog.Trace.Util
                     (segment.Length == 32 && IsAGuid(segment, "N")) ||
                     (segment.Length == 36 && IsAGuid(segment, "D")))
                 {
-                    segment = UrlIdPlaceholder;
+                    segment = "?";
                 }
 
                 sb.Append(segment);
@@ -84,10 +122,62 @@ namespace Datadog.Trace.Util
                 previousIndex = index + 1;
             }
 
-            return sb.ToString();
+            return StringBuilderCache.GetStringAndRelease(sb);
+#endif
         }
 
         [MethodImpl(MethodImplOptions.NoInlining)]
         private static bool IsAGuid(string segment, string format) => Guid.TryParseExact(segment, format, out _);
+
+#if NETCOREAPP
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private static bool IsAGuid(ReadOnlySpan<char> segment, string format) => Guid.TryParseExact(segment, format, out _);
+#endif
+
+        /// <summary>
+        /// Provide a cached reusable instance of stringbuilder per thread.
+        /// </summary>
+        /// <remarks>
+        /// Based on https://source.dot.net/#System.Private.CoreLib/StringBuilderCache.cs,a6dbe82674916ac0
+        /// </remarks>
+        internal static class StringBuilderCache
+        {
+            internal const int MaxBuilderSize = 360;
+
+            [ThreadStatic]
+            private static StringBuilder _cachedInstance;
+
+            public static StringBuilder Acquire(int capacity)
+            {
+                if (capacity <= MaxBuilderSize)
+                {
+                    StringBuilder sb = _cachedInstance;
+                    if (sb != null)
+                    {
+                        // Avoid stringbuilder block fragmentation by getting a new StringBuilder
+                        // when the requested size is larger than the current capacity
+                        if (capacity <= sb.Capacity)
+                        {
+                            _cachedInstance = null;
+                            sb.Clear();
+                            return sb;
+                        }
+                    }
+                }
+
+                return new StringBuilder(capacity);
+            }
+
+            public static string GetStringAndRelease(StringBuilder sb)
+            {
+                string result = sb.ToString();
+                if (sb.Capacity <= MaxBuilderSize)
+                {
+                    _cachedInstance = sb;
+                }
+
+                return result;
+            }
+        }
     }
 }